Split NewSavingsGoalDialog spec into NewSavingsGoalDialog-tests/ directory
The original spec file was 407 lines, exceeding the 360-line split
threshold. Split into two domain-focused files:
- NewSavingsGoalDialogView.spec.tsx (view/presentational tests)
- NewSavingsGoalDialog.spec.tsx (integration/container tests)

All 25 tests pass unchanged.
